# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: MÉRITO API — Talent identity & validation platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all public projects
      parameters:
        - name: userId
          in: query
          schema:
            type: integer
        - name: skill
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"

    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a project by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found

    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdate"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"

    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── SAR EXPERIENCES ───────────────────────────────────────────────────────
  /sar:
    get:
      operationId: listSarExperiences
      tags: [sar]
      summary: List SAR experiences for the authenticated user
      parameters:
        - name: userId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: SAR experiences
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SarExperience"

    post:
      operationId: createSarExperience
      tags: [sar]
      summary: Create a SAR experience
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SarInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SarExperience"
  /sar/{id}:
    get:
      operationId: getSarExperience
      tags: [sar]
      summary: Get a SAR experience
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: SAR experience
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SarExperience"

    patch:
      operationId: updateSarExperience
      tags: [sar]
      summary: Update a SAR experience
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SarUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SarExperience"

    delete:
      operationId: deleteSarExperience
      tags: [sar]
      summary: Delete a SAR experience
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── SKILLS ────────────────────────────────────────────────────────────────
  /skills:
    get:
      operationId: listSkills
      tags: [skills]
      summary: List skills for a user
      parameters:
        - name: userId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Skills
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Skill"

    post:
      operationId: addSkill
      tags: [skills]
      summary: Add a skill
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SkillInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Skill"
  /skills/{id}:
    delete:
      operationId: deleteSkill
      tags: [skills]
      summary: Remove a skill
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── VALIDATIONS ───────────────────────────────────────────────────────────
  /validations:
    get:
      operationId: listValidations
      tags: [validations]
      summary: List validations received by a user
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Validations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Validation"

    post:
      operationId: createValidation
      tags: [validations]
      summary: Create a validation for a student
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ValidationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Validation"

  # ── MICROPROJECTS ─────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Project:
      type: object
      required: [id, userId, title, description, problem, solution, impact, role, technologies, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        title:
          type: string
        description:
          type: string
        problem:
          type: string
        solution:
          type: string
        impact:
          type: string
        role:
          type: string
        technologies:
          type: array
          items:
            type: string
        imageUrl:
          type: ["string", "null"]
        githubUrl:
          type: ["string", "null"]
        figmaUrl:
          type: ["string", "null"]
        demoUrl:
          type: ["string", "null"]
        videoUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    ProjectInput:
      type: object
      required: [title, description, problem, solution, impact, role, technologies]
      properties:
        title:
          type: string
          minLength: 2
        description:
          type: string
        problem:
          type: string
        solution:
          type: string
        impact:
          type: string
        role:
          type: string
        technologies:
          type: array
          items:
            type: string
        imageUrl:
          type: string
        githubUrl:
          type: string
        figmaUrl:
          type: string
        demoUrl:
          type: string
        videoUrl:
          type: string
    ProjectUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        problem:
          type: string
        solution:
          type: string
        impact:
          type: string
        role:
          type: string
        technologies:
          type: array
          items:
            type: string
        imageUrl:
          type: string
        githubUrl:
          type: string
        figmaUrl:
          type: string
        demoUrl:
          type: string
        videoUrl:
          type: string

    # ── SAR SCHEMAS ───────────────────────────────────────────────────────
    SarExperience:
      type: object
      required: [id, userId, situacion, accion, resultado, softSkills, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        situacion:
          type: string
        accion:
          type: string
        resultado:
          type: string
        softSkills:
          type: array
          items:
            type: string
        context:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    SarInput:
      type: object
      required: [situacion, accion, resultado, softSkills]
      properties:
        situacion:
          type: string
          minLength: 10
        accion:
          type: string
          minLength: 10
        resultado:
          type: string
          minLength: 10
        softSkills:
          type: array
          items:
            type: string
        context:
          type: string
    SarUpdate:
      type: object
      properties:
        situacion:
          type: string
        accion:
          type: string
        resultado:
          type: string
        softSkills:
          type: array
          items:
            type: string
        context:
          type: string

    # ── SKILL SCHEMAS ─────────────────────────────────────────────────────
    Skill:
      type: object
      required: [id, userId, name, category, level]
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        category:
          type: string
          enum: [technical, soft, tool, language]
        level:
          type: string
          enum: [beginner, intermediate, advanced, expert]
    SkillInput:
      type: object
      required: [name, category, level]
      properties:
        name:
          type: string
          minLength: 1
        category:
          type: string
          enum: [technical, soft, tool, language]
        level:
          type: string
          enum: [beginner, intermediate, advanced, expert]

    # ── VALIDATION SCHEMAS ────────────────────────────────────────────────
    Validation:
      type: object
      required: [id, validatedUserId, validatorId, skillName, message, createdAt]
      properties:
        id:
          type: integer
        validatedUserId:
          type: integer
        validatorId:
          type: integer
        validatorName:
          type: ["string", "null"]
        skillName:
          type: string
        message:
          type: string
        rating:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
    ValidationInput:
      type: object
      required: [validatedUserId, skillName, message]
      properties:
        validatedUserId:
          type: integer
        skillName:
          type: string
        message:
          type: string
          minLength: 10
        rating:
          type: integer
          minimum: 1
          maximum: 5

    # ── MICROPROJECT SCHEMAS ──────────────────────────────────────────────
...