# lib/api-zod/src/generated/types/sarInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MÉRITO API — Talent identity & validation platform
 * OpenAPI spec version: 0.1.0
 */

export interface SarInput {
  /** @minLength 10 */
  situacion: string;
  /** @minLength 10 */
  accion: string;
  /** @minLength 10 */
  resultado: string;
  softSkills: string[];
  context?: string;
}